Refactor environment variable usage and utility functions across the application
- Replaced instances of `getBaseUrl` with `nextPublicBaseUrl` for consistent environment variable access.
- Moved utility functions like `slugify` and `formatPrice` to `lib/format.ts` for better organization.
- Introduced `scrollToSection` function in `lib/dom.ts` for improved DOM manipulation.
- Updated API routes to utilize the new environment variable structure, enhancing maintainability and clarity.
