+++ components/menu/SimpleMenu/SimpleMenuProduct.tsx
@@ -1,4 +1,4 @@
-import { formatPrice } from "@/lib/utils";
+import { formatPrice } from "@/lib/format";
 import { FC } from "react";
 
 interface PublicProduct {
@@ -22,21 +22,15 @@ const SimpleMenuProduct: FC<Props> = ({ product, currency }) => {
       className={`flex  justify-between py-2 border-b border-gray-200/50 dark:border-gray-200/50 last:border-b-0 ${!product.is_available ? "opacity-60" : ""}`}
     >
       <div className="flex-1">
-        <div
-          className={`flex items-center ${!product.is_available ? "text-muted-foreground line-through" : "text-[#8B1538] dark:text-[#A61E4D]"}`}
-        >
+        <div className={`flex items-center ${!product.is_available ? "text-muted-foreground line-through" : "text-[#8B1538] dark:text-[#A61E4D]"}`}>
           <div className="flex justify-between items-center w-full">
             <p className="font-semibold flex-shrink-0">{product.name}</p>
             <div className="border-b border-dashed border-gray-200/50 dark:border-gray-200  w-full mx-12 py-1"></div>
-            {!!product.price && product.price > 0 && currency && (
-              <p className="flex-shrink-0 block">{formatPrice(product.price, currency)}</p>
-            )}
+            {!!product.price && product.price > 0 && currency && <p className="flex-shrink-0 block">{formatPrice(product.price, currency)}</p>}
           </div>
         </div>
         {product.description && (
-          <p className="text-gray-500 dark:text-gray-400 text-xs mt-1 leading-relaxed font-normal italic">
-            {product.description}
-          </p>
+          <p className="text-gray-500 dark:text-gray-400 text-xs mt-1 leading-relaxed font-normal italic">{product.description}</p>
         )}
       </div>
     </div>
