+++ lib/dom.ts
@@ -1,0 +1,7 @@
+export const scrollToSection = (sectionId: string) => {
+  const element = document.getElementById(sectionId);
+
+  if (element) {
+    element.scrollIntoView({ behavior: "smooth" });
+  }
+};
