+++ lib/env.ts
@@ -67,7 +67,3 @@ export const getSupabaseEnvironment = () => {
 export const nextPublicBaseUrl = getNextPublicBaseUrl();
 
 export const supabaseConfig = getSupabaseEnvironment();
-
-console.log(nextPublicBaseUrl);
-
-console.log(supabaseConfig);
